+++ deployment/helm/skaha/CHANGELOG.md
@@ -1,4 +1,7 @@
-# CHANGELOG for Skaha User Session API (Chart 0.6.0)
+# CHANGELOG for Skaha User Session API (Chart 0.7.0)
+
+## 2024.09.20 (0.7.0)
+- Allow setting nodeAffinity values for proper scheduling.
 
 ## 2024.09.20 (0.6.0)
 - Feature to allow mounting volumes into user sessions.